requirements.txt
aiogram>=3.0.0
asyncpg
python-dotenv
sqlalchemy[asyncio]
alembic